BUG! There's a bug.
I was investigating the population fraction numbers
and some of them were greater than 1. I clicked on a cell
and watched the evolution of it's state with time. It appeared
that one of the cells cycled through a loop going from 9-<1
and over again. That's incorrect.

TODO: Inspect the model and make changes where appropriate.